[editor] Adding skeleton of Executor py API
